backend/requirements.txt
# FastAPI e dependências
fastapi==0.108.0
uvicorn[standard]==0.25.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Pydantic
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# MinIO/S3
boto3==1.34.18
minio==7.2.3

# Utils
python-dateutil==2.8.2
pytz==2023.3

# Excel/CSV
openpyxl==3.1.2
pandas==2.1.4

# Logs estruturados
structlog==24.1.0

# Jobs/Cron
apscheduler==3.10.4

# Fiscal - NF-e
zeep==4.2.1
lxml==4.9.4
cryptography==41.0.7
signxml==3.2.1
pyOpenSSL==24.0.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
reportlab==4.1.0
python-barcode==0.15.1
Pillow==10.2.0
